requirements.txt
streamlit
openai
python-dotenv
PyMuPDF
python-docx
pyttsx3
llama-cpp-python
pydub
requests
